requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-filter==24.2
markdown==3.6

# Database
psycopg2-binary==2.9.9

uwsgi==2.0.26

# Checker
pycodestyle
pylint==3.2.*
pylint-django==2.5.5
